backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
requests
beautifulsoup4
PyPDF2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==0.21.1
google-generativeai
sqlalchemy==2.0.23
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.3.0
sqlite3